SAML support changes
